+++ resources_test/run_predict_modality.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# run prior to running this script:
+# bin/viash_build -q 'common|predict_modality'
+
 set -ex
 
 # get the root of the directory
@@ -16,22 +19,24 @@ mkdir -p `dirname $out_file`
 $target_dir/predict_modality_datasets/censor_dataset/censor_dataset \
   --input_mod1 ${in_file}.output_rna.h5ad \
   --input_mod2 ${in_file}.output_mod2.h5ad \
-  --output_mod1 ${out_file}.mod1.h5ad \
-  --output_mod2 ${out_file}.mod2.h5ad \
-  --output_solution ${out_file}.solution.h5ad
+  --output_train_mod1 ${out_file}.train_mod1.h5ad \
+  --output_train_mod2 ${out_file}.train_mod2.h5ad \
+  --output_test_mod1 ${out_file}.test_mod1.h5ad \
+  --output_test_mod2 ${out_file}.test_mod2.h5ad
   
 $target_dir/predict_modality_methods/baseline_randomforest/baseline_randomforest \
-  --input_mod1 ${out_file}.mod1.h5ad \
-  --input_mod2 ${out_file}.mod2.h5ad \
+  --input_train_mod1 ${out_file}.train_mod1.h5ad \
+  --input_train_mod2 ${out_file}.train_mod2.h5ad \
+  --input_test_mod1 ${out_file}.test_mod1.h5ad \
   --output ${out_file}.prediction.h5ad
   
 $target_dir/predict_modality_metrics/calculate_cor/calculate_cor \
   --input_prediction ${out_file}.prediction.h5ad \
-  --input_solution ${out_file}.solution.h5ad \
+  --input_solution ${out_file}.test_mod2.h5ad \
   --output ${out_file}.scores.h5ad
 
 $target_dir/common/extract_scores/extract_scores \
   --input ${out_file}.scores.h5ad \
-  --metric_meta src/predict_modality/metrics/calculate_cor/metric_meta.tsv \
+  --metric_meta src/predict_modality/metrics/calculate_cor/metric_meta_calculate_cor.tsv \
   --output ${out_file}.scores.tsv \
   --summary ${out_file}.summary.tsv