+++ src/common/workflows/generate_datasets/run.sh
@@ -6,7 +6,6 @@ REPO_ROOT=$(git rev-parse --show-toplevel)
 # ensure that the command below is run from the root of the repository
 cd "$REPO_ROOT"
 
-
 export NXF_VER=21.04.1
 
 bin/nextflow \