+++ src/predict_modality/methods/baseline_knearestneighbors/script.R
@@ -6,32 +6,38 @@ library(Matrix, warn.conflicts = FALSE, quietly = TRUE)
 
 ## VIASH START
 par <- list(
-  input_mod1 = "resources_test/predict_modality/pbmc_1k_protein_v3.mod1.h5ad",
-  input_mod2 = "resources_test/predict_modality/pbmc_1k_protein_v3.mod2.h5ad",
+  input_train_mod1 = "resources_test/predict_modality/test_resource.train_mod1.h5ad",
+  input_test_mod1 = "resources_test/predict_modality/test_resource.test_mod1.h5ad",
+  input_train_mod2 = "resources_test/predict_modality/test_resource.train_mod2.h5ad",
   output = "output.h5ad",
-  n_pcs = 4L
+  n_pcs = 4L,
+  n_neighbors = 3
 )
 ## VIASH END
 
 cat("Reading h5ad files\n")
-ad1 <- anndata::read_h5ad(par$input_mod1)
-ad2 <- anndata::read_h5ad(par$input_mod2)
+ad1_train <- anndata::read_h5ad(par$input_train_mod1)
+ad1_test <- anndata::read_h5ad(par$input_test_mod1)
+ad2 <- anndata::read_h5ad(par$input_train_mod2)
 
 cat("Performing DR on the mod1 values\n")
+# LMDS is more efficient than regular MDS because
+# it does not compure a square distance matrix.
 dr <- lmds::lmds(
-  ad1$X, 
+  rbind(ad1_train$X, ad1_test$X),
   ndim = par$n_pcs,
   distance_method = par$distance_method
 )
 
-dr_train <- dr[ad1$obs$group == "train",]
+ix <- seq_len(nrow(ad1_train))
+dr_train <- dr[ix, , drop = FALSE]
+dr_test <- dr[-ix, , drop = FALSE]
 responses_train <- ad2$X
-dr_test <- dr[ad1$obs$group == "test",]
 
 cat("Predicting for each column in modality 2\n")
 preds <- apply(responses_train, 2, function(yi) {
   FNN::knn.reg(
-    train = dr_train, 
+    train = dr_train,
     test = dr_test,
     y = yi,
     k = par$n_neighbors
@@ -39,8 +45,9 @@ preds <- apply(responses_train, 2, function(yi) {
 })
 
 cat("Creating outputs object\n")
+# store prediction as a sparse matrix
 prediction <- Matrix::Matrix(
-  preds, 
+  preds,
   sparse = TRUE,
   dimnames = list(rownames(dr_test), colnames(ad2))
 )
@@ -48,7 +55,7 @@ prediction <- Matrix::Matrix(
 out <- anndata::AnnData(
   X = prediction,
   uns = list(
-    dataset_id = ad1$uns[["dataset_id"]],
+    dataset_id = ad1_train$uns[["dataset_id"]],
     method_id = "baseline_knearestneighbors"
   )
 )